feat(tools): add cancellable spawn tasks
