+++ crates/tools/src/spawn_agent_tasks.rs
@@ -12,6 +12,7 @@ use {
 use crate::{error::Error, params::str_param};
 
 const DEFAULT_TASK_TTL_HOURS: i64 = 24;
+const DEFAULT_STALE_RUNNING_TTL_HOURS: i64 = 48;
 const CLEANUP_INTERVAL_SECS: i64 = 60;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -19,6 +20,7 @@ pub enum SpawnTaskStatus {
     Running,
     Completed,
     Failed,
+    Cancelled,
 }
 
 impl SpawnTaskStatus {
@@ -27,6 +29,43 @@ impl SpawnTaskStatus {
             Self::Running => "running",
             Self::Completed => "completed",
             Self::Failed => "failed",
+            Self::Cancelled => "cancelled",
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct SpawnTaskRetention {
+    completed_ttl: Duration,
+    failed_ttl: Duration,
+    running_ttl: Duration,
+    cleanup_interval: Duration,
+}
+
+impl Default for SpawnTaskRetention {
+    fn default() -> Self {
+        Self {
+            completed_ttl: Duration::hours(DEFAULT_TASK_TTL_HOURS),
+            failed_ttl: Duration::hours(DEFAULT_TASK_TTL_HOURS),
+            running_ttl: Duration::hours(DEFAULT_STALE_RUNNING_TTL_HOURS),
+            cleanup_interval: Duration::seconds(CLEANUP_INTERVAL_SECS),
+        }
+    }
+}
+
+impl SpawnTaskRetention {
+    #[must_use]
+    pub fn new(
+        completed_ttl: Duration,
+        failed_ttl: Duration,
+        running_ttl: Duration,
+        cleanup_interval: Duration,
+    ) -> Self {
+        Self {
+            completed_ttl,
+            failed_ttl,
+            running_ttl,
+            cleanup_interval,
         }
     }
 }
@@ -56,13 +95,19 @@ pub struct SpawnTask {
 }
 
 impl SpawnTask {
-    fn is_expired(&self, now: OffsetDateTime, ttl: Duration) -> bool {
-        match self.finished_at {
+    fn is_expired(&self, now: OffsetDateTime, retention: &SpawnTaskRetention) -> bool {
+        match (self.status.clone(), self.finished_at) {
             // Completed/failed tasks expire after one TTL from completion.
-            Some(finished_at) => finished_at + ttl <= now,
-            // Running tasks that never completed get a grace period of 2× TTL
-            // from their start time before being reaped as stale.
-            None => self.started_at + ttl + ttl <= now,
+            (SpawnTaskStatus::Completed, Some(finished_at)) => {
+                finished_at + retention.completed_ttl <= now
+            },
+            (SpawnTaskStatus::Failed | SpawnTaskStatus::Cancelled, Some(finished_at)) => {
+                finished_at + retention.failed_ttl <= now
+            },
+            // Running tasks that never completed get a separate stale-running TTL.
+            (SpawnTaskStatus::Running, _) => self.started_at + retention.running_ttl <= now,
+            // Malformed terminal records without finished_at should not be retained forever.
+            (_, None) => self.started_at + retention.failed_ttl <= now,
         }
     }
 
@@ -103,35 +148,30 @@ impl SpawnTask {
 #[derive(Debug)]
 pub struct SpawnTaskStore {
     tasks: RwLock<HashMap<String, SpawnTask>>,
-    ttl: Duration,
-    cleanup_interval: Duration,
+    abort_handles: RwLock<HashMap<String, futures::future::AbortHandle>>,
+    retention: SpawnTaskRetention,
     last_cleanup: std::sync::Mutex<Option<OffsetDateTime>>,
 }
 
 impl Default for SpawnTaskStore {
     fn default() -> Self {
-        Self::new(Duration::hours(DEFAULT_TASK_TTL_HOURS))
+        Self::new(SpawnTaskRetention::default())
     }
 }
 
 impl SpawnTaskStore {
-    pub fn new(ttl: Duration) -> Self {
+    pub fn new(retention: SpawnTaskRetention) -> Self {
         Self {
             tasks: RwLock::new(HashMap::new()),
-            ttl,
-            cleanup_interval: Duration::seconds(CLEANUP_INTERVAL_SECS),
+            abort_handles: RwLock::new(HashMap::new()),
+            retention,
             last_cleanup: std::sync::Mutex::new(None),
         }
     }
 
     #[cfg(test)]
-    fn with_cleanup_interval(ttl: Duration, cleanup_interval: Duration) -> Self {
-        Self {
-            tasks: RwLock::new(HashMap::new()),
-            ttl,
-            cleanup_interval,
-            last_cleanup: std::sync::Mutex::new(None),
-        }
+    fn with_retention(retention: SpawnTaskRetention) -> Self {
+        Self::new(retention)
     }
 
     #[tracing::instrument(skip(self, task), fields(model = %model))]
@@ -141,6 +181,7 @@ impl SpawnTaskStore {
         session_key: Option<String>,
         model: String,
         preset: Option<String>,
+        abort_handle: futures::future::AbortHandle,
     ) -> SpawnTask {
         let entry = SpawnTask {
             id: Uuid::new_v4().to_string(),
@@ -160,12 +201,20 @@ impl SpawnTaskStore {
             .write()
             .await
             .insert(entry.id.clone(), entry.clone());
+        self.abort_handles
+            .write()
+            .await
+            .insert(entry.id.clone(), abort_handle);
         entry
     }
 
     #[tracing::instrument(skip(self, update))]
     pub async fn complete(&self, id: &str, update: SpawnTaskUpdate) {
+        self.abort_handles.write().await.remove(id);
         if let Some(task) = self.tasks.write().await.get_mut(id) {
+            if task.status == SpawnTaskStatus::Cancelled {
+                return;
+            }
             task.status = if update.error.is_some() {
                 SpawnTaskStatus::Failed
             } else {
@@ -177,6 +226,33 @@ impl SpawnTaskStore {
             task.tool_calls_made = update.tool_calls_made;
             task.error = update.error;
         }
+    }
+
+    pub async fn cancel(&self, id: &str, session_key: Option<&str>) -> crate::Result<Value> {
+        let now = OffsetDateTime::now_utc();
+        self.cleanup_expired(now).await;
+        let mut tasks = self.tasks.write().await;
+        let task = tasks
+            .get_mut(id)
+            .ok_or_else(|| Error::message(format!("spawn task not found: {id}")))?;
+        task.assert_access(session_key)?;
+        if task.status != SpawnTaskStatus::Running {
+            let mut value = task.status_json(now);
+            value["cancelled"] = false.into();
+            return Ok(value);
+        }
+
+        task.status = SpawnTaskStatus::Cancelled;
+        task.finished_at = Some(now);
+        task.error = Some("cancelled by caller".to_string());
+        let mut value = task.status_json(now);
+        value["cancelled"] = true.into();
+        drop(tasks);
+
+        if let Some(handle) = self.abort_handles.write().await.remove(id) {
+            handle.abort();
+        }
+        Ok(value)
     }
 
     #[tracing::instrument(skip(self))]
@@ -220,8 +296,24 @@ impl SpawnTaskStore {
         }
         let mut tasks = self.tasks.write().await;
         let before = tasks.len();
-        tasks.retain(|_, task| !task.is_expired(now, self.ttl));
+        let mut expired_ids = Vec::new();
+        tasks.retain(|id, task| {
+            let keep = !task.is_expired(now, &self.retention);
+            if !keep {
+                expired_ids.push(id.clone());
+            }
+            keep
+        });
         let expired = before - tasks.len();
+        drop(tasks);
+        if !expired_ids.is_empty() {
+            let mut abort_handles = self.abort_handles.write().await;
+            for id in expired_ids {
+                if let Some(handle) = abort_handles.remove(&id) {
+                    handle.abort();
+                }
+            }
+        }
 
         #[cfg(feature = "metrics")]
         if expired > 0 {
@@ -237,7 +329,7 @@ impl SpawnTaskStore {
             .last_cleanup
             .lock()
             .unwrap_or_else(|poisoned| poisoned.into_inner());
-        if last_cleanup.is_some_and(|last| last + self.cleanup_interval > now) {
+        if last_cleanup.is_some_and(|last| last + self.retention.cleanup_interval > now) {
             return false;
         }
         *last_cleanup = Some(now);
@@ -367,22 +459,73 @@ impl AgentTool for SpawnListTool {
     }
 }
 
+#[derive(Clone)]
+pub struct SpawnCancelTool {
+    store: Arc<SpawnTaskStore>,
+}
+
+impl SpawnCancelTool {
+    pub fn new(store: Arc<SpawnTaskStore>) -> Self {
+        Self { store }
+    }
+}
+
+#[async_trait]
+impl AgentTool for SpawnCancelTool {
+    fn name(&self) -> &str {
+        "cancel_spawn"
+    }
+
+    fn description(&self) -> &str {
+        "Cancel a running non-blocking spawn_agent task by task_id."
+    }
+
+    fn parameters_schema(&self) -> Value {
+        serde_json::json!({
+            "type": "object",
+            "properties": {
+                "task_id": {
+                    "type": "string",
+                    "description": "Task ID returned by spawn_agent with nonblocking=true."
+                }
+            },
+            "required": ["task_id"]
+        })
+    }
+
+    #[tracing::instrument(skip(self, params))]
+    async fn execute(&self, params: Value) -> anyhow::Result<Value> {
+        let id = str_param(&params, "task_id")
+            .ok_or_else(|| Error::message("missing required parameter: task_id"))?;
+        let session_key = str_param(&params, "_session_key");
+        Ok(self.store.cancel(id, session_key).await?)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
 
+    fn abort_handle() -> futures::future::AbortHandle {
+        let (handle, _registration) = futures::future::AbortHandle::new_pair();
+        handle
+    }
+
     #[tokio::test]
     async fn stale_running_tasks_are_cleaned_up() {
-        let store = SpawnTaskStore::with_cleanup_interval(
+        let store = SpawnTaskStore::with_retention(SpawnTaskRetention::new(
+            Duration::hours(1),
+            Duration::hours(1),
             Duration::milliseconds(1),
             Duration::milliseconds(0),
-        );
+        ));
         let task = store
             .insert_running(
                 "zombie task".to_string(),
                 None,
                 "mock-model".to_string(),
                 None,
+                abort_handle(),
             )
             .await;
         tokio::time::sleep(std::time::Duration::from_millis(3)).await;
@@ -395,13 +538,19 @@ mod tests {
 
     #[tokio::test]
     async fn cleanup_is_amortized_between_polls() {
-        let store = SpawnTaskStore::with_cleanup_interval(Duration::hours(1), Duration::minutes(1));
+        let store = SpawnTaskStore::with_retention(SpawnTaskRetention::new(
+            Duration::hours(1),
+            Duration::hours(1),
+            Duration::hours(1),
+            Duration::minutes(1),
+        ));
         let task = store
             .insert_running(
                 "active task".to_string(),
                 None,
                 "mock-model".to_string(),
                 None,
+                abort_handle(),
             )
             .await;
 
@@ -437,6 +586,7 @@ mod tests {
                 Some("session-1".to_string()),
                 "model".to_string(),
                 None,
+                abort_handle(),
             )
             .await;
         store
@@ -445,6 +595,7 @@ mod tests {
                 Some("session-2".to_string()),
                 "model".to_string(),
                 None,
+                abort_handle(),
             )
             .await;
         store
@@ -453,6 +604,7 @@ mod tests {
                 Some("session-1".to_string()),
                 "model".to_string(),
                 None,
+                abort_handle(),
             )
             .await;
 
@@ -471,5 +623,50 @@ mod tests {
         let tool = SpawnListTool::new(Arc::new(SpawnTaskStore::default()));
         let schema = tool.parameters_schema();
         assert!(schema.get("required").is_none());
+    }
+
+    #[tokio::test]
+    async fn cancel_marks_running_task_cancelled() {
+        let store = SpawnTaskStore::default();
+        let task = store
+            .insert_running(
+                "cancel me".to_string(),
+                Some("session-1".to_string()),
+                "model".to_string(),
+                None,
+                abort_handle(),
+            )
+            .await;
+
+        let cancelled = store.cancel(&task.id, Some("session-1")).await.unwrap();
+
+        assert_eq!(cancelled["status"], "cancelled");
+        assert_eq!(cancelled["cancelled"], true);
+        assert_eq!(cancelled["error"], "cancelled by caller");
+    }
+
+    #[tokio::test]
+    async fn cancel_enforces_session_key() {
+        let store = SpawnTaskStore::default();
+        let task = store
+            .insert_running(
+                "private task".to_string(),
+                Some("session-1".to_string()),
+                "model".to_string(),
+                None,
+                abort_handle(),
+            )
+            .await;
+
+        let denied = store.cancel(&task.id, Some("session-2")).await;
+
+        assert!(denied.is_err());
+        assert!(denied.unwrap_err().to_string().contains("access denied"));
+    }
+
+    #[test]
+    fn cancel_spawn_tool_requires_task_id() {
+        let tool = SpawnCancelTool::new(Arc::new(SpawnTaskStore::default()));
+        assert_eq!(tool.parameters_schema()["required"][0], "task_id");
     }
 }
